Separate tag and metadata tests, add more tests
